requirements.txt
attrs==23.2.0
aws-cdk-lib==2.146.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
black==24.8.0
cattrs==23.2.3
click==8.1.7
constructs==10.3.0
exceptiongroup==1.2.2
importlib_resources==6.4.0
jsii==1.102.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.12.2